requirements.txt
beautifulsoup4==4.9.3
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
Flask==1.1.2
html5lib==1.1
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.3
MarkupSafe==1.1.1
psycopg2-binary==2.8.6
pyTelegramBotAPI==3.7.6
python-decouple==3.4
python-dotenv==0.15.0
requests==2.25.1
six==1.15.0
soupsieve==2.2
urllib3==1.26.3
webencodings==0.5.1
Werkzeug==1.0.1
